Replaced comments with functioning code in document generation
Added dependency imports, uncommented and filled functionality code in the order_doc function in the hrdepartment_app/models.py file. This allows the conversion of html to docx to be saved as a sub-document which is integrated into the main document. Also added file remove command to delete the subdoc file once it's content has been embedded.

Implemented to improve the way documents are generated, particularly enriching the description detail section and ensuring temporary files are appropriately cleaned up.