Refactored job query and removed print statements
Replaced a print statement in the hrdepartment_app/views.py with a comment as it was used for debugging purposes and no longer necessary. In administration_app/views.py, refactored a queryset for Job objects to a more specific query for ReportCard objects related to the current user's division affiliation. This streamlines the data processing and avoids unnecessary iterations over the Job objects. This was done to optimize performance and enhance readability.