Add QR code generation and display across different pages
QR code generation logic has been added to the administration_app context_processors.py file making use of the qrcode library. This QR code is based on the absolute URI of the request, saved in a relevant path and the link is sent to the template page for display to the user. This change allows users to scan the QR code to directly access a specific page on their mobile devices. This commit also includes a change in the left_menu.html of library_app to accommodate and display the newly generated QR code. Necessary changes were made in the settings.py file to include the new context processor function across the project.