Ensure consistent indents for condition checks
Made several adjustments to indents for if-else conditions all across the hrdepartment_app/views.py file. These changes aim to improve code readability, as the previous indents were inconsistent and did not adhere strictly to Python's PEP 8 style guide. Along with this, also added functionality to display a link to the parent document in the Job Description update page and updated the handling of 'previous_document' field in the 'DocumentsJobDescription' model.