Changed date difference and added a notice in report form
Increased the permissible past period from 60 to 90 days in forms.py to give users more flexibility. Also added a notice in reportcard_form.html to inform users that they don't need to manually enter data if they have a business trip or an official outing. Further, extended the 'yesterday' definition from 3 to 7 days in a couple of instances in forms.py to cater for longer weekends or holidays.