Add 'creation_retroactively' field in Official Memo Model
A new BooleanField 'creation_retroactively' was added to the 'OfficialMemo' model in 'hrdepartment_app'. The 'creation_retroactively' field with default value 'False' is used to indicate if a document was introduced retroactively. Also, this field is integrated into the form and migrations files accordingly. A typo was also corrected in 'hrdepartment_app/views.py' and data validation is improved in 'customer_app/customers_util.py'.