Improve HR memo form and data validation
This change includes the introduction of the 'creation_retroactively' field in the OfficialMemo model. This new field allows users to indicate if a memo document was introduced retroactively, adding more flexibility and accuracy to data entry. Updates made to the form and migrations files integrate this change throughout the application. In addition to this, a typo was fixed in 'hrdepartment_app/views.py'. Enhancements were also made to the data validation process in 'customer_app/customers_util.py' for improved reliability.